Revert "Purge notifications for configured days of retention"
